Update apptx.yaml
Changed line 820, from
        image: image-registry.openshift-image-registry.svc:5000/sock-shop/user
to
	quay.io/daniel_casali/usvc-user

Updated documentation as below
Important Note

There has been a recent change to an opensource component that is required by the sock-shop application. On Sunday 10th November, I noticed that the openzipkin component had changed, causing the build of the container to fail this . The change must have taken place since the end of October, when I last recorded the demonstration. The current output from the podman build command is below. 

For this reason, please now skip to chapter 5. Building-a-.NET-Service-from-a-UBI-image

I will modify the application yaml file to pull the prebuilt Go User Application container from Quay.io
-----------------------------------